[SCFToCalyx] [StaticLogicToCalyx] Share passes.
